Remove use of uint from particle random system
uint (and bitwise operations) aren't supported by our minimum spec, and
they're going to be removed from public SkSL. For now, convert the
random generator to a good-enough chaotic sequence of high-frequency
sine waves.

If/when the interpreter (and particles) are converted to the newer skvm
backend, it will be straightforward to support custom intrinsics that
emit skvm instructions directly into the builder, and re-introduce a
better integer-based PRNG, without requiring SkSL language support.

Bug: skia:11093

+++ modules/particles/include/SkParticleEffect.h
@@ -64,7 +64,7 @@ public:
     //   float  spin  = 0;               // Angular velocity, in (radians / second)
     //   float4 color = { 1, 1, 1, 1 };  // RGBA color
     //   float  frame = 0;               // Normalized sprite index for multi-frame drawables
-    //   uint   seed  = 0;               // Random seed, used with rand() (see below)
+    //   float  seed  = 0;               // Random value, used with rand() (see below)
     // };
     //
     // Particle functions are defined in fParticleCode, and get a mutable Particle struct, as well
@@ -80,12 +80,12 @@ public:
     //   float  spin;
     //   float4 color;
     //   float  frame;
-    //   uint   seed;
+    //   float  seed;
     // };
     //
-    // All functions have access to a global function named 'rand'. It takes a uint seed value,
-    // which it uses and updates (using a linear congruential RNG). It returns a random floating
-    // point value in [0, 1]. Typical usage is to pass the particle or effect's seed value to rand.
+    // All functions have access to a global function named 'rand'. It takes a float seed value,
+    // which it uses and updates (using a PRNG). It returns a random floating point value in [0, 1].
+    // Typical usage is to pass the particle or effect's seed value to rand.
     // For particle functions, the seed is rewound after each update, so calls to 'rand(p.seed)'
     // will return consistent values from one update to the next.
     //
@@ -141,7 +141,7 @@ public:
 
     // Start playing this effect, specifying initial values for the emitter's properties
     void start(double now, bool looping, SkPoint position, SkVector heading, float scale,
-               SkVector velocity, float spin, SkColor4f color, float frame, uint32_t seed);
+               SkVector velocity, float spin, SkColor4f color, float frame, float seed);
 
     // Start playing this effect, with default values for the emitter's properties
     void start(double now, bool looping) {
@@ -153,7 +153,7 @@ public:
                     0.0f,                        // spin
                     { 1.0f, 1.0f, 1.0f, 1.0f },  // color
                     0.0f,                        // sprite frame
-                    0);                          // seed
+                    0.0f);                       // seed
     }
 
     void update(double now);
@@ -229,7 +229,7 @@ private:
         float     fSpin;
         SkColor4f fColor;
         float     fFrame;
-        uint32_t  fRandom;
+        float     fRandom;
     };
     EffectState fState;
